[analyzer] Fix an assertion fail in hash generation.
In case the (uniqueing) location of the diagnostic is in a line that only
contains whitespaces there was an assertion fail during issue hash generation.

Unfortunately I am unable to reproduce this error with the built in checkers,
so no there is no failing test case with this patch. It would be possible to
write a debug checker for that purpuse but it does not worth the effort.

Differential Revision: http://reviews.llvm.org/D18210

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@264851 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit a26e4e76c2abe29a074fac69ae56575a5d050333)

apple-llvm-split-commit: 46411ca70cc1925c4ff1e6a6db6d4a95baa70fd3
apple-llvm-split-dir: clang/