[DependenceAnalysis] Check if result of getConstantPart is null
A seg-fault occurs due to a reference of a null pointer, which is
the value returned by getConstantPart. This function returns
null if the constant part is not found. The code that calls this
function needs to check for the null return value.

Differential Revision: http://reviews.llvm.org/D18718

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@265319 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 3188534401388fa57dd125df768f90158fe06629)

apple-llvm-split-commit: 62056e3831dbdea8e1c3f9358365686ed4d55714
apple-llvm-split-dir: llvm/