[tsan] Remove long sleeps from fork tests
On one of our testing machines, we're running the tests under heavy load, and especially in the fork-based TSan tests, we're seeing timeouts when a test uses sleep(10), assuming that calling fork() on another thread will finish sooner than that. This patch removes a timeout and makes another one longer.

Differential Revision: http://reviews.llvm.org/D18476



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@265666 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 7462b2d8b8b3193d98a6741f7832fb31b96f673a
apple-llvm-split-dir: compiler-rt/