Linker: Always pass RF_IgnoreMissingLocals; NFC
This is a cleanup after clarifying the meaning of RF_IgnoreMissingLocals
in r265628 and truly limiting it to locals in r265768.

This should have no functionality change, since the only context that
the flag has an effect is when we could hit function-local Value and
Metadata, and we were already passing it in those contexts.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@265831 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 6900b6439360da3012ba97dfa05788c18f47f212)

apple-llvm-split-commit: 12ca2b434d73b9408b07ca3f8b60af19d513ff23
apple-llvm-split-dir: llvm/