[CMake] If set we should pass LLVM_VERSION_INFO into config.h
Autoconf used to support setting LLVM_VERSION_INFO and there is some code filtered around llvm in Support/CommandLine.cpp and LTO/LTOCodeGenerator.cpp that uses it if it is set.

We also shouldn't be explicitly setting it as a define on llvm-shlib. It is pointless there because there is no code using it in llvm-shlib, and it is better to have it as part of the generated config.h so that it is available everywhere.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@267490 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 7901ae5c6ddf84dbac88458d24e9c6ed2948373f
apple-llvm-split-dir: llvm/