[analyzer] Fix crash in ObjCGenericsChecker
Fix a crash in the generics checker where DynamicTypePropagation tries
to get the superclass of a root class.

This is a spot-fix for a deeper issue where the checker makes assumptions
that may not hold about subtyping between the symbolically-tracked type of
a value and the compile-time types of a cast on that value.

I've added a TODO to address the underlying issue.

rdar://problem/26086914

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@269227 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit da822bb6244dfc205c1a7f617226ef0d26b352c9)

apple-llvm-split-commit: 3e4242e46bd4ef849dc530ece76207be8c9f4da9
apple-llvm-split-dir: clang/