[sanitizer] On OS X, exit the forked process gracefully when login_tty fails
We're using forkpty to spawn the atos symbolizer. In some cases, login_tty (part of forkpty) can fail due to security measures (sandboxing). In this case, we should exit with a status code instead of completely crashing the spawned process. Even processing a failed CHECK() is problematic here, because we're post-fork and pre-exec where a lot of things don't work (for multithreaded processes, for OS X GUI apps, etc.).

Differential Revision: http://reviews.llvm.org/D20048



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@269289 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: cfc9db796ab0df32ceccc65e8dbd6f03100b233c
apple-llvm-split-dir: compiler-rt/