Separate the Verifier into an analysis and a transformation pass and allow the transformation to strip invalid debug info.
This patch separates the Verifier into an analysis and a transformation
pass, with the transformation pass optionally stripping malformed
debug info.

The problem I'm trying to solve with this sequence of patches is that
historically we've done a really bad job at verifying debug info. We want
to be able to make the verifier stricter without having to worry about
breaking bitcode compatibility with existing producers. For example, we
don't necessarily want IR produced by an older version of clang to be
rejected by an LTO link just because of malformed debug info, and rather
provide an option to strip it. Note that merely outdated (but well-formed)
debug info would continue to be auto-upgraded in this scenario.

http://reviews.llvm.org/D19988
rdar://problem/25818489

This reapplies r268937 without modifications.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@268966 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit f68fb38629ac39d7ffced4d519069ed8c367fcbd)

Reapplied with backport changes included this time.

# Conflicts:
#	include/llvm/IR/Verifier.h
#	lib/Passes/PassRegistry.def
#	unittests/IR/VerifierTest.cpp

apple-llvm-split-commit: 4387c2546c49933b067b57439f3dd497a4a705c3
apple-llvm-split-dir: llvm/