Removed the m_decl_objects map from ClangASTContext.
m_decl_objects is problematic because it assumes that each VarDecl has a unique
variable associated with it.  This is not the case in inline contexts.

Also the information in this map can be reconstructed very easily without
maintaining the map.  The rest of the testsuite passes with this cange, and I've
added a testcase covering the inline contexts affected by this.

<rdar://problem/26278502>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@270474 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: fb807f34f866ae646a8f41483137f75e4d12d4a9
apple-llvm-split-dir: lldb/